+++ src/main/java/com/example/Library/repo/RentalRepository.java
@@ -3,6 +3,6 @@ package com.example.Library.repo;
 import com.example.Library.model.Rental;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalRepository extends JpaRepository<Rental, Integer> {
+public interface RentalRepository extends JpaRepository<Rental, Long> {
 
 }
